#24 chore: merge CommonLogic into CalendarLogic
